add roadmap to v0.1.0
The initial approach of calling Get() directly on our API object is too
simplistic to be of much use. We need, at a bare minimum, to be able to create
a number of Endpoint objects, and then create (and modify) a Query on an
Endpoint (with AddHeader(), Add?Query?()) and finally calling it (with GET(),
POST(), etc) once it is fully set up.