revise Ident() code
It turns out the error identified in `issue 0003` was not actually an error
at all; it was a side effect of a glitch during editing that removed the wrong
line of code. I noticed that and fixed it before merging -- but it did suggest
to me that it would be worth testing for both the default name AND a blank name.