add Status to *Result; convert Body to string
We can now pass the StatusCode back to the client and allow them to decide
whether it represents an error or not.

Additionally, we have changed the Body value from []byte to string. This may
be a better approach, or it may just be because I'm not used to working with
[]byte values -- ie, personal preference. If this turns out to be a bad choice
I'm sure we'll figure it out.