refactor!: remove string conversions
Mostly because I come from other languages, I tend to be more comfortable
working with a string than with a []byte slice, so my code tends to convert back
to string whenever possible. This is unnecessary effort that is probably wasting
time and memory.

BREAKING CHANGE: Response.Body changed from string to []byte