feat: ignore possible changes to existing ratelimiter
Once a rateLimiter has been set on an endPoint, it is fixed.

Given that the code to create the EP (and hence the associated
SetRateLimit() call) may be called numerous times, we need to
give some thought to how to change the Rate once it has been
set. It may require quite a bit of hand-waving...